Introduce `suspenders:testing:minitest` and `suspenders:testing:rspec`
