+++ lib/generators/templates/testing/rspec/action_mailer.rb
