+++ lib/generators/templates/testing/rspec/chromedriver.rb
