+++ lib/generators/templates/testing/rspec/i18n.rb
